Migrate request endpoints into submodule
